+++ bugsnag-android-performance/src/main/kotlin/com/bugsnag/android/performance/NetworkRequestInfo.kt
@@ -1,9 +1,9 @@
 package com.bugsnag.android.performance
 
-class NetworkRequestInfo(
+public class NetworkRequestInfo(
     /**
      * The URL that will be reported in this network request's span.
      * If null, no span will be created.
      */
-    var url: String? = null,
+    public var url: String? = null,
 )