+++ bugsnag-android-performance/src/main/kotlin/com/bugsnag/android/performance/internal/InternalDebug.kt
@@ -1,18 +1,21 @@
 package com.bugsnag.android.performance.internal
 
+import androidx.annotation.RestrictTo
+
 /**
  * Warning: Everything in here is to support internal testing, and is subject to change
  * without notice.
  * DO NOT USE!
  */
-object InternalDebug {
-    var spanBatchSizeSendTriggerPoint = 100
+@RestrictTo(RestrictTo.Scope.LIBRARY_GROUP_PREFIX)
+public object InternalDebug {
+    public var spanBatchSizeSendTriggerPoint: Int = 100
 
     /**
      * The maximum amount of time the worker thread will sleep without a `wake()`
      */
-    var workerSleepMs = 30000L
+    public var workerSleepMs: Long = 30000L
 
-    var dropSpansOlderThanMs = 24L * 60 * 60 * 1000
-    var pValueExpireAfterMs = 24 * 60 * 60 * 1000
+    public var dropSpansOlderThanMs: Long = 24L * 60 * 60 * 1000
+    public var pValueExpireAfterMs: Int = 24 * 60 * 60 * 1000
 }