+++ bugsnag-android-performance/src/main/kotlin/com/bugsnag/android/performance/internal/NetworkType.kt
@@ -1,8 +1,8 @@
 package com.bugsnag.android.performance.internal
 
-enum class NetworkType(
+internal enum class NetworkType(
     @JvmField
-    internal val otelName: String
+    internal val otelName: String,
 ) {
     WIFI("wifi"),
     WIRED("wired"),