+++ bugsnag-android-performance/src/main/kotlin/com/bugsnag/android/performance/internal/SpanCategory.kt
@@ -1,6 +1,9 @@
 package com.bugsnag.android.performance.internal
 
-enum class SpanCategory(val category: String?) {
+import androidx.annotation.RestrictTo
+
+@RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
+public enum class SpanCategory(public val category: String?) {
     CUSTOM(null),
     VIEW_LOAD("view_load"),
     VIEW_LOAD_PHASE("view_load_phase"),