+++ bugsnag-android-performance/src/main/kotlin/com/bugsnag/android/performance/internal/SpanProcessor.kt
@@ -1,7 +1,9 @@
 package com.bugsnag.android.performance.internal
 
+import androidx.annotation.RestrictTo
 import com.bugsnag.android.performance.Span
 
-fun interface SpanProcessor {
-    fun onEnd(span: Span)
+@RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
+public fun interface SpanProcessor {
+    public fun onEnd(span: Span)
 }