+++ bugsnag-plugin-android-performance-appcompat/src/main/kotlin/com/bugsnag/android/performance/AppCompatModule.kt
@@ -3,7 +3,7 @@ package com.bugsnag.android.performance
 import com.bugsnag.android.performance.internal.InstrumentedAppState
 import com.bugsnag.android.performance.internal.Module
 
-class AppCompatModule : Module {
+public class AppCompatModule : Module {
     private lateinit var instrumentedAppState: InstrumentedAppState
     private lateinit var fragmentActivityLifecycleCallbacks: FragmentActivityLifecycleCallbacks
     override fun load(instrumentedAppState: InstrumentedAppState) {