+++ bugsnag-plugin-android-performance-appcompat/src/main/kotlin/com/bugsnag/android/performance/FragmentActivityLifecycleCallbacks.kt
@@ -8,11 +8,12 @@ import androidx.fragment.app.Fragment
 import androidx.fragment.app.FragmentActivity
 import androidx.fragment.app.FragmentManager
 import androidx.fragment.app.FragmentManager.FragmentLifecycleCallbacks
+import com.bugsnag.android.performance.internal.AutoInstrumentationCache
 import com.bugsnag.android.performance.internal.SpanFactory
 import com.bugsnag.android.performance.internal.SpanTracker
 import com.bugsnag.android.performance.internal.ViewLoadPhase
 
-class FragmentActivityLifecycleCallbacks(
+public class FragmentActivityLifecycleCallbacks(
     private val spanTracker: SpanTracker,
     private val spanFactory: SpanFactory,
     private val autoInstrumentationCache: AutoInstrumentationCache,
@@ -85,10 +86,10 @@ class FragmentActivityLifecycleCallbacks(
         spanTracker.endAllSpans(f)
     }
 
-    override fun onActivityStarted(activity: Activity) = Unit
-    override fun onActivityResumed(activity: Activity) = Unit
-    override fun onActivityPaused(activity: Activity) = Unit
-    override fun onActivityStopped(activity: Activity) = Unit
-    override fun onActivitySaveInstanceState(activity: Activity, outState: Bundle) = Unit
-    override fun onActivityDestroyed(activity: Activity) = Unit
+    override fun onActivityStarted(activity: Activity): Unit = Unit
+    override fun onActivityResumed(activity: Activity): Unit = Unit
+    override fun onActivityPaused(activity: Activity): Unit = Unit
+    override fun onActivityStopped(activity: Activity): Unit = Unit
+    override fun onActivitySaveInstanceState(activity: Activity, outState: Bundle): Unit = Unit
+    override fun onActivityDestroyed(activity: Activity): Unit = Unit
 }